move any and every
