+++ addon/array/index.js
@@ -9,6 +9,8 @@ export { default as find } from './find';
 export { default as first } from './first';
 export { default as includes } from './includes';
 export { default as indexOf } from './index-of';
+export { default as isAny } from './is-any';
+export { default as isEvery } from './is-every';
 export { default as lastIndexOf } from './last-index-of';
 export { default as last } from './last';
 export { default as length } from './length';
