+++ addon/array/is-every.js
@@ -1,7 +1,7 @@
-import { normalizeArray } from './utils';
+import { normalizeArray } from '../utils';
 
 export default function(...keys) {
   return normalizeArray(keys, {}, (array, key, value) => {
-    return array.isAny(key, value);
+    return array.isEvery(key, value);
   });
 }
